+++ run.py
@@ -6,6 +6,7 @@ from flask_cache import Cache
 from flask_sockets import Sockets
 import time
 import ujson
+import logging
 
 app = Flask(__name__)
 cache = Cache(app, config={'CACHE_TYPE': 'simple', 'CACHE_DEFAULT_TIMEOUT': 0})
@@ -18,15 +19,16 @@ sockets = Sockets(app)
 @sockets.route('/getdata')
 def feed(websocket):
     message = ujson.loads(websocket.receive())
-    measurements = test_measurements()
-    print(message)
+    with open("epiviz.json", "r") as config_file:
+        measurements = ujson.loads(config_file)
+    logging.info(message)
     data = message['data']
     start = data['start']
     end = data['end']
     chromosome = data['chr']
     gene_name = data['gene']
     seqID = message['seq']
-    print ("parameters")
+    logging.info ("parameters")
     key = chromosome + '-' + str(start) + '-' + str(end)
     cached = cache.get(key)
     if cached:
@@ -35,11 +37,11 @@ def feed(websocket):
         return
     results = computational_request(start, end, chromosome, gene_name, measurements=measurements)
     cache_results = []
-    print (results)
+    logging.info (results)
     for result in results:
-        print ("send back!")
-        print (time.time())
-        print ("\n")
+        logging.info ("send back!")
+        logging.info (time.time())
+        logging.info ("\n")
         # emit('returned_results', result)
         cache_results.extend(result)
         websocket.send(ujson.dumps(result))
@@ -49,145 +51,12 @@ def feed(websocket):
     websocket.send(ujson.dumps(seqID))
 
 
-# just for testing purposes
-def test_measurements(expression=True, block=True, methylation=True):
-    measurements = []
-    gene_types = ['breast___normal', 'breast___tumor', 'colon___normal',
-                  'colon___tumor', 'lung___normal', 'lung___tumor',
-                  'thyroid___normal', 'thyroid___tumor']
-
-    gene_names = ['breast_normal', 'breast_tumor', 'colon_normal',
-                  'colon_tumor', 'lung_normal', 'lung_tumor',
-                  'thyroid_normal', 'thyroid_tumor']
-
-    tissue_types = ['breast', 'colon', 'thyroid', 'lung']
-
-    methylation_types = ['breast_normal', 'breast_cancer', 'colon_normal',
-                         'colon_cancer', 'lung_normal', 'lung_cancer',
-                         'thyroid_normal', 'thyroid_cancer']
-    if expression:
-        for gene_type, gene_name in zip(gene_types, gene_names):
-            measurements.append({
-                "id": gene_type,
-                "name": 'Expression ' + gene_name,
-                "type": "feature",
-                "datasourceId": "gene_expression_barcode_subtype",
-                "datasourceGroup": "gene_expression_barcode_subtype",
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "scatterplot",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-    if block:
-        for tissue_type in tissue_types:
-            measurements.append({
-                "id": 'timp2014_' + tissue_type + '_blocks',
-                "name": tissue_type + ' blocks',
-                "type": "feature",
-                "datasourceId": 'timp2014_' + tissue_type + '_blocks',
-                "datasourceGroup": 'timp2014_' + tissue_type + '_blocks',
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "block",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-    if methylation:
-        # for methylation difference
-        for tissue_type in tissue_types:
-            measurements.append({
-                "id": tissue_type,
-                "name": 'Collapsed Methylation Diff ' + tissue_type,
-                "type": "feature",
-                "datasourceId": 'timp2014_collapsed_diff',
-                "datasourceGroup": 'timp2014_collapsed_diff',
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "line",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-        for methylation_type in methylation_types:
-            measurements.append({
-                "id": methylation_type,
-                "name": ' Average Probe level Meth ' + methylation_type,
-                "type": "feature",
-                "datasourceId": 'timp2014_probelevel_beta',
-                "datasourceGroup": 'timp2014_probelevel_beta',
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "line",
-                "annotation": None,
-                "metadata": []
-            })
-
-    return measurements
-
-
-# road map measurements
-def roadmap_measurements(expression=True, block=True, methylation=True):
-    measurements = []
-    gene_types = ['breast___normal', 'breast___tumor', 'colon___normal',
-                  'colon___tumor', 'lung___normal', 'lung___tumor',
-                  'thyroid___normal', 'thyroid___tumor']
-
-    tissue_types = ['breast', 'colon', 'thyroid', 'lung']
-    if expression:
-        for gene_type in gene_types:
-            measurements.append({
-                "id": gene_type,
-                "name": gene_type,
-                "type": "feature",
-                "datasourceId": "gene_expression_barcode_subtype",
-                "datasourceGroup": "gene_expression_barcode_subtype",
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "scatterplot",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-    if block:
-        for tissue_type in tissue_types:
-            measurements.append({
-                "id": 'timp2014_' + tissue_type + '_blocks',
-                "name": 'timp2014_' + tissue_type + '_blocks',
-                "type": "feature",
-                "datasourceId": 'timp2014_' + tissue_type + '_blocks',
-                "datasourceGroup": 'timp2014_' + tissue_type + '_blocks',
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "block",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-    if methylation:
-        for tissue_type in tissue_types:
-            measurements.append({
-                "id": tissue_type,
-                "name": tissue_type,
-                "type": "feature",
-                "datasourceId": 'timp2014_collapsed_diff',
-                "datasourceGroup": 'timp2014_collapsed_diff',
-                "dataprovider": "umd",
-                "formula": None,
-                "defaultChartType": "line",
-                "annotation": None,
-                "metadata": ["probe"]
-            })
-
-    return measurements
-
-
 if __name__ == "__main__":
     from gevent import pywsgi
     from geventwebsocket.handler import WebSocketHandler
     server = pywsgi.WSGIServer(('', 5001), app, handler_class=WebSocketHandler)
-    print(sockets.route('/getdata'))
-    print ("Server Starts!")
+    # data = ujson.dumps(test_measurements())
+    # with open('epiviz.json', 'w') as outfile:
+    #     json.dump(test_measurements(), outfile)
+    logging.info("Server Starts!")
     server.serve_forever()
